Docker/Data-Ingestion-Service/requirements.txt
kafka-python
pandas
pyspark
pymongo
pymongo[srv]
schedule
requests